监控箱删除申请单时应该弹出是否确认删除的对话框

+++ workflow/client/views/instance/instance_button.coffee
@@ -100,7 +100,16 @@ Template.instance_button.events
         InstanceManager.saveIns();
 
     'click #instance_remove': (event)->
-        InstanceManager.deleteIns();
+        swal {   
+            title: t("Are you sure?"),    
+            type: "warning",   
+            showCancelButton: true,  
+            cancelButtonText: t('Cancel'), 
+            confirmButtonColor: "#DD6B55",   
+            confirmButtonText: t('OK'),   
+            closeOnConfirm: true 
+        }, () ->  
+            InstanceManager.deleteIns()
 
     'click #instance_submit': (event)->
         InstanceManager.checkFormValue();
